Register hotkeys only on save
