+++ LightBulb/Internal/Extensions.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.ComponentModel;
 using System.Text;
 
 namespace LightBulb.Internal
@@ -26,18 +25,5 @@ namespace LightBulb.Internal
                 : dateTime.AddDays(-1).AtTimeOfDay(timeOfDay);
 
         public static void OpenInBrowser(this Uri uri) => ProcessEx.StartShellExecute(uri.ToString());
-
-        public static void Bind<TSource>(this TSource target, EventHandler<PropertyChangedEventArgs> handler)
-            where TSource : class, INotifyPropertyChanged
-        {
-            target.PropertyChanged += (sender, args) => handler(target, args);
-        }
-
-        public static void BindAndInvoke<TSource>(this TSource target, EventHandler<PropertyChangedEventArgs> handler)
-            where TSource : class, INotifyPropertyChanged
-        {
-            target.Bind(handler);
-            handler.Invoke(target, new PropertyChangedEventArgs(null));
-        }
     }
 }
