+++ LightBulb/ViewModels/Dialogs/SettingsViewModel.cs
@@ -1,22 +1,30 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using LightBulb.Messages;
+using LightBulb.Models;
 using LightBulb.Services;
 using LightBulb.ViewModels.Components;
 using LightBulb.ViewModels.Framework;
+using Stylet;
 
 namespace LightBulb.ViewModels.Dialogs
 {
     public class SettingsViewModel : DialogScreen
     {
+        private readonly IEventAggregator _eventAggregator;
         private readonly SettingsService _settingsService;
+        private readonly SystemService _systemService;
 
         public IReadOnlyList<ISettingsTabViewModel> Tabs { get; }
 
         public ISettingsTabViewModel ActiveTab { get; private set; }
 
-        public SettingsViewModel(SettingsService settingsService, IEnumerable<ISettingsTabViewModel> tabs)
+        public SettingsViewModel(IEventAggregator eventAggregator, SettingsService settingsService, SystemService systemService,
+            IEnumerable<ISettingsTabViewModel> tabs)
         {
             _settingsService = settingsService;
+            _eventAggregator = eventAggregator;
+            _systemService = systemService;
 
             Tabs = tabs.OrderBy(t => t.Order).ToArray();
 
@@ -42,6 +50,7 @@ namespace LightBulb.ViewModels.Dialogs
         public void Save()
         {
             _settingsService.Save();
+            RegisterHotKeys();
             Close(true);
         }
 
@@ -49,6 +58,16 @@ namespace LightBulb.ViewModels.Dialogs
         {
             _settingsService.Load();
             Close(false);
+        }
+
+        public void RegisterHotKeys()
+        {
+            _systemService.UnregisterAllHotKeys();
+
+            if (_settingsService.ToggleHotKey != HotKey.None)
+            {
+                _systemService.RegisterHotKey(_settingsService.ToggleHotKey, () => _eventAggregator.Publish(new ToggleIsEnabledMessage()));
+            }
         }
     }
 }
